auge: @grant none: close #117, localstore fixes
We don't use GM_functions in recent Greasemonkey anymore
due to "@grant none" in the header.
They are only used on platforms with old GM (< 1.0)
or GM emulations.
We add dummies when neither localstore nor GM_functions are available,
so we print only one (visible) warning.
We do print a warning for every key we try to set if we are not able
due to a reached quota and possibly a private mode.
Every private mode I have seen still allows localstorage,
but might delete it afterwards.

git-svn-id: http://svn.kraehen.org/tools/auge/trunk@1009 e0e46c49-be69-4f5a-ad62-21024a331aea